requirements.txt

# pip install torch==2.2.0 torchvision==0.17.0 torchaudio==2.2.0 --index-url https://download.pytorch.org/whl/cu121
# pip install --upgrade torch==2.3.0 torchvision==0.18.0 torchaudio==2.3.0 --index-url https://download.pytorch.org/whl/cu121

lightning==2.2.0
jsonargparse[signatures]>=4.27.7
tensorboard
tensorboardx
albumentations==1.4.4
omegaconf
einops
requests
transformers==4.37.2
lpips
matplotlib
cmaps
natsort
pandas
xarray
h5netcdf
seaborn
properscoring
boto3
botocore
wandb==0.15.0
termcolor
tensorboard
segmentation_models_pytorch
cfgrib
pygrib
ninja==1.11.1
submitit==1.5.1
moviepy==1.0.3
dict_recursive_update==1.0.1
packaging
ninja
eccodes==1.2.0